requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cloudinary==1.29.0
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.1
PyJWT==1.7.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==5.3.0
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
cryptography==37.0.2
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.3.1
django-rest-passwordreset==1.2.1
django-webpush==0.3.4
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
http-ece==1.1.0
idna==3.3
Pillow==8.4.0
psycopg2==2.9.1
py-vapid==1.8.2
pycparser==2.21
PyJWT==1.7.1
pytz==2021.3
pywebpush==1.9.4
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==5.3.0
